+++ CHANGES.md
@@ -1,13 +1,19 @@
 Changlog for Opevario Compman
 =============================
 
-0.3 (unreleased)
+0.4 (unreleased)
 ----------------
 
+- Allow to select xcsoar profiles where to install selected competition files
 - Correctly reference waypoint and airspace files in xcsoar profile
-
 - Integrate with [openvario-shell](https://github.com/kedder/openvario-shell)
 
+
+0.3 (2020-04-03)
+----------------
+
+- Copy competition files to XCSoar home directory. This would allow to backup
+  and move xcsoar profile between devices without breaking configuration.
 
 
 0.2 (2020-03-10)